Add more info for zypper install phase
Adds global option `-vv` and install option `--details`,
to add a little more information to the logging.

Install option --download-in-advance is required to enable
package File Conflicts to be reported before any installation
occurs.  The docker image provided zypp.conf leaves this
configuration unset, so it is enabled in this change.